+++ src/app/app.module.ts
@@ -4,28 +4,24 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 
-import { ProductComponent } from './product-list/product/product.component';
-import { ProductListComponent } from './product-list/product-list.component';
-
-import { CartComponent } from './cart-list/cart/cart.component';
-import { CartListComponent } from './cart-list/cart-list.component';
-
 import { ProductsService } from './product-list/products.service';
 import { CartService } from './cart-list/cart.service';
 
-
-
+import { CartModule } from './cart-list/cart.module';
+import { SharedModule } from './shared/shared.module';
+import { ProductsModule } from './product-list/products.module';
+import { CoreModule } from './core/core.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartComponent,
-    CartListComponent
   ],
   imports: [
-    BrowserModule
+    BrowserModule,
+    SharedModule,
+    CoreModule,
+    CartModule,
+    ProductsModule
   ],
   providers: [
     ProductsService,