+++ src/app/cart-list/cart.service.ts
@@ -1,23 +1,26 @@
 import { Injectable } from '@angular/core';
 
-import { ProductModel } from '../models/product.model';
 import { ProductsService } from '../product-list/products.service';
+import { CartItem } from '../core/interfaces/CartItem';
+import { CartItemModel } from '../models/cartItem.model';
 
 @Injectable()
 export class CartService {
-  private cartItems: Array<ProductModel> = [];
+  private cartItems: Array<CartItem> = [];
 
   constructor(private productsService: ProductsService) {
 
   }
 
   addItem(id) {
-    const isExist = this.cartItems.find(item => item.id === id);
+    const cartItem = this.cartItems.find(item => item.product.id === id);
 
-    if (!isExist) {
+    if (!cartItem) {
       const item = this.productsService.getProductById(id);
 
-      this.cartItems.push(item);
+      this.cartItems.push(new CartItemModel(item, 1));
+    } else {
+      cartItem.quantity += 1;
     }
   }
 
@@ -25,4 +28,8 @@ export class CartService {
     return this.cartItems;
   }
 
+  // TODO: show sum of Product
+  // getCartItemsPrice() {
+  //   return this.cartItems.reduce((sum, {product}) => sum + product.price, 0);
+  // }
 }