+++ src/app/cart-list/cart/cart.component.ts
@@ -1,6 +1,6 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
-import { ProductModel } from '../../models/product.model';
+import { CartItem } from '../../core/interfaces/CartItem';
 
 @Component({
   selector: 'app-cart',
@@ -8,16 +8,18 @@ import { ProductModel } from '../../models/product.model';
   styleUrls: ['./cart.component.css']
 })
 export class CartComponent implements OnInit {
-  @Input() cartList: Array<ProductModel>;
+  @Input() cartList: Array<CartItem>;
+  @Output() removeProduct: EventEmitter<any> = new EventEmitter();
 
+  // TODO - issues/1
   cartColumns = [
     {key: 'id', title: 'ID'},
     {key: 'name', title: 'Name'},
     {key: 'description', title: 'Description'},
     {key: 'price', title: 'Price'},
+    {key: 'quantity', title: 'Quantity'},
     {key: 'category', title: 'Category'},
-    {key: 'types', title: 'Types(s)'},
-    {key: 'isAvailable', title: 'isAvailable'},
+    {key: 'types', title: 'Types(s)'}
   ];
 
   constructor() {
@@ -25,6 +27,14 @@ export class CartComponent implements OnInit {
 
   ngOnInit() {
 
+  }
+
+  /**
+   * Remove product.
+   * @param idx
+   */
+  onRemove(idx): void {
+    this.removeProduct.emit(idx);
   }
 
 }