+++ src/app/product-list/product-list.component.ts
@@ -1,4 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+
+import { ProductsService } from './products.service';
+import { ProductModel } from '../models/product.model';
+import { CartService } from '../cart-list/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -6,9 +10,22 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./product-list.component.css']
 })
 export class ProductListComponent implements OnInit {
-  constructor() { }
+  @Output() buy: EventEmitter<any> = new EventEmitter<any>();
 
-  ngOnInit() {
+  products: Array<ProductModel>;
+
+  constructor(private productsService: ProductsService,
+              private cartService: CartService) {
   }
 
+  ngOnInit() {
+    this.products = this.productsService.getProducts();
+  }
+
+  onBuy(data) {
+    this.buy.emit(data);
+    this.cartService.addItem(data);
+    console.log('product ' + data + ' added to cart');
+  }
 }
+