+++ src/app/product-list/products.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ProductListComponent } from './product-list.component';
+import { ProductComponent } from './product/product.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [
+    ProductComponent,
+    ProductListComponent
+  ],
+  exports: [
+    ProductComponent,
+    ProductListComponent
+  ]
+})
+export class ProductsModule { }